+++ dmbc/src/currency/transactions/builders/wallet.rs
@@ -29,7 +29,7 @@ impl Builder {
         Builder { balance, ..self }
     }
 
-    pub fn add_asset(self, name: &str, amount: u32) -> Self {
+    pub fn add_asset(self, name: &str, amount: u64) -> Self {
         assert!(self.public_key.is_some());
         let id = AssetId::new(name, self.public_key.as_ref().unwrap()).unwrap();
         let asset = Asset::new(id, amount);
